Move merge_parallel_block_layers to block.py
