+++ src/tqec/compile/blocks/__init__.py
@@ -34,6 +34,7 @@ See :mod:`tqec.compile.blocks.layers` for more details.
 """
 
 from .block import Block as Block
+from .block import merge_parallel_block_layers as merge_parallel_block_layers
 from .enums import SpatialBlockBorder as SpatialBlockBorder
 from .enums import TemporalBlockBorder as TemporalBlockBorder
 from .layers import BaseComposedLayer as BaseComposedLayer
@@ -43,5 +44,4 @@ from .layers import PlaquetteLayer as PlaquetteLayer
 from .layers import RawCircuitLayer as RawCircuitLayer
 from .layers import RepeatedLayer as RepeatedLayer
 from .layers import SequencedLayers as SequencedLayers
-from .layers import merge_parallel_block_layers as merge_parallel_block_layers
 from .positioning import LayoutPosition2D as LayoutPosition2D
