+++ src/tqec/compile/blocks/layers/__init__.py
@@ -5,4 +5,3 @@ from .atomic import RawCircuitLayer as RawCircuitLayer
 from .composed import BaseComposedLayer as BaseComposedLayer
 from .composed import RepeatedLayer as RepeatedLayer
 from .composed import SequencedLayers as SequencedLayers
-from .merge import merge_parallel_block_layers as merge_parallel_block_layers
